+++ src/components/Button/Button.tsx
@@ -6,14 +6,14 @@ export class Button
   extends React.Component<ButtonProps> {
 
   render() {
-    const { title, type, disable, onclick } = this.props;
+    const { title, type, disable, onClick } = this.props;
     const style = this.stylingButton(type, disable);
 
     return (
       <button
         className={style}
         disabled={disable}
-        onClick={onclick}
+        onClick={onClick}
       >
         {title}
       </button>