+++ src/pages/SelectFriendPage/SelectFriendPageTypes.ts
@@ -3,12 +3,26 @@ export interface SelectFriendState {
   friends: Friend[];
   selectedFriendList: Friend[];
   selectedFriend: number;
+  totalPrice: number;
+  items: Item[];
+  vat: number;
+  vatStatus: boolean;
+  vatPrice: number;
   totalBillPrice: number;
+  serviceCharge: number;
+  serviceChargeStatus: boolean;
+  serviceChargePrice: number;
 }
 
 export interface Friend {
+  userId: string;
   profilePic: string;
   displayName: string;
   isSelect: boolean;
   owner: boolean;
+}
+
+export interface Item {
+  detail: string;
+  price: number;
 }