+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [V0.0.17] - 2019-06-28
+## Updated
+- Added history state to pass between page.
+- Added removeIsSelect function.
+- Added addIsSelect function.
+
 ## [V0.0.16] - 2019-06-26
 ## Updated
 - Updated stepper component css style.